requirements.txt
streamlit>=1.35.0
pandas>=2.1.0
plotly>=5.20.0
google-api-python-client>=2.120.0
google-auth>=2.28.0
openpyxl>=3.1.2
pyarrow>=15.0.0
python-dotenv>=1.0.1
numpy>=1.26.0
